Inline opaque wrapper regression fixture
